Implement NCI-3 offline GorXu cognition
