+++ src/grox/reasoning/__init__.py
@@ -1,10 +1,12 @@
-from .base import CognitiveUsage, ReasoningProvider, ReasoningError
-from .contracts import MissionInterpretation, StrategyOption
+from .base import CognitiveUsage, ConversationalReasoningProvider, ReasoningProvider, ReasoningError
+from .contracts import AssistantResponse, MissionInterpretation, StrategyOption
 from .factory import build_reasoner_from_env
 from .session import SessionReasoningProvider
 
 __all__ = [
     "CognitiveUsage",
+    "ConversationalReasoningProvider",
+    "AssistantResponse",
     "ReasoningProvider",
     "ReasoningError",
     "MissionInterpretation",
